% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snps-data.R
\docType{data}
\name{snps}
\alias{snps}
\title{Dataset containing details on simulated SNPs}
\format{A data frame with 20 rows and 12 variables:
\describe{
  \item{positions}{location of the SNP on the chromosome}
  \item{distances}{genetic distance of the SNP from the start of the chromosome}
  \item{chrom_1}{allele of the SNP on the first example chromosome}
  \item{chrom_2}{allele of the SNP on the second example chromosome}
  \item{chrom_3}{allele of the SNP on the third example chromosome}
  \item{chrom_4}{allele of the SNP on the fourth example chromosome}
  \item{chrom_5}{allele of the SNP on the fifth example chromosome}
  \item{chrom_6}{allele of the SNP on the sixth example chromosome}
  \item{chrom_7}{allele of the SNP on the seventh example chromosome}
  \item{chrom_8}{allele of the SNP on the eighth example chromosome}
  \item{chrom_9}{allele of the SNP on the ninth example chromosome}
  \item{chrom_10}{allele of the SNP on the tenth example chromosome}
}}
\usage{
data(snps)
}
\description{
A dataset containing the positions, genetic distances and
alleles for 20 SNPs, across 10 simulated chromosomes.
}
\keyword{datasets}
