% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zalpha.R
\name{Zalpha}
\alias{Zalpha}
\title{Runs the Zalpha function}
\usage{
Zalpha(pos, x, ws, minRandL = 4, minRL = 25)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{x}{A matrix of SNP values. Columns represent chromosomes, rows are SNP locations. Hence, the number of rows should equal the length of the `pos` vector.}

\item{ws}{The window size which the `Zalpha` statistic will calculate over. This should be on the same scale as the `pos` vector.}

\item{minRandL}{Minimum number of SNPs in each set R and L for the statistic to be calculated. Default is 4.}

\item{minRL}{Minimum value for the product of the set sizes for R and L. Default is 25.}
}
\value{
A vector of Zalpha values
}
\description{
Returns a Zalpha value for each SNP location supplied to the function.
}
