% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zalpha_expected.R
\name{Zalpha_expected}
\alias{Zalpha_expected}
\title{Runs the Zalpha function on the expected r squared values for the region}
\usage{
Zalpha_expected(pos, cM, ws, LDprofile_cM_bins, LDprofile_rsq,
  minRandL = 4, minRL = 25, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{cM}{A mnumeric vector of cM distances. This should be the same length as \code{pos}.}

\item{ws}{The window size which the \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} statistic will be calculated over. This should be on the same scale as the \code{pos} vector.}

\item{LDprofile_cM_bins}{A numeric vector containing the lower bound of the bins used in the LD profile. These should be of equal size.}

\item{LDprofile_rsq}{A numeric vector containing the expected \eqn{r^2}{r^2} values for the corresponding bin in the LD profile. Must be between 0 and 1.}

\item{minRandL}{Minimum number of SNPs in each set R and L for the statistic to be calculated. Default is 4.}

\item{minRL}{Minimum value for the product of the set sizes for R and L. Default is 25.}

\item{X}{Optional. Specify a region of the chromosome to calculate \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} for in the format \code{c(startposition, endposition)}. The start position and the end position should be within the extremes of the positions given in the \code{pos} vector. If not supplied, the function will calculate \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} for every SNP in the \code{pos} vector.}
}
\value{
A list containing the SNP positions and the \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} values for those SNPs
}
\description{
Returns a \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} value for each SNP location supplied to the function, based on
the expected \eqn{r^2} values given an LD profile and cM distances.
For more information about the \eqn{Z_{\alpha}^{E[r^2]}}{Zalpha} statistic please see Jacobs (2016).
The \eqn{Z_{\alpha}^{E[r^2]}} statistic is defined as:
\deqn{{Z_{\alpha}^{E[r^2]}}=\frac{{|L| \choose 2}^{-1}\sum_{i,j \in L}E[r^2_{i,j}] + {|R| \choose 2}^{-1}\sum_{i,j \in L}E[r^2_{i,j}]}{2}}
where \code{|L|} and \code{|R|} are the number of SNPs to the left and right of the current locus within the given window \code{ws},
and \eqn{E[r^2]}{E[r^2]} is equal to the expected squared correlation between a pair of SNPs, given an LD profile.
}
\details{
The LD profile describes the expected correlation between SNPs at a given cM distance, generated using simulations or
real data. Care should be taken to utilise an LD profile which is representative of the population in question. The LD
profile should consist of evenly-sized bins of distances (for example 0.00001 cM per bin), where the value given is the (inclusive) lower
bound of the bin.
}
\references{
Jacobs, G.S., T.J. Sluckin, and T. Kivisild, \emph{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps.} Genetics, 2016. \strong{203}(4): p. 1807
}
