% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zbeta_BetaCDF.R
\name{Zbeta_BetaCDF}
\alias{Zbeta_BetaCDF}
\title{Runs the Zbeta function using a cumulative beta distribution function on the r-squared values for the region}
\usage{
Zbeta_BetaCDF(pos, x, dist, ws, LDprofile_bins, LDprofile_Beta_a,
  LDprofile_Beta_b, minRandL = 4, minRL = 25, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{x}{A matrix of SNP values. Columns represent chromosomes, rows are SNP locations. Hence, the number of rows should equal the length of the \code{pos} vector. SNPs should all be biallelic.}

\item{dist}{A numeric vector of genetic distances (e.g. cM, LDU). This should be the same length as \code{pos}.}

\item{ws}{The window size which the \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} statistic will be calculated over. This should be on the same scale as the \code{pos} vector.}

\item{LDprofile_bins}{A numeric vector containing the lower bound of the bins used in the LD profile. These should be of equal size.}

\item{LDprofile_Beta_a}{A numeric vector containing the first estimated Beta parameter for the corresponding bin in the LD profile.}

\item{LDprofile_Beta_b}{A numeric vector containing the second estimated Beta parameter for the corresponding bin in the LD profile.}

\item{minRandL}{Minimum number of SNPs in each set R and L for the statistic to be calculated. Default is 4.}

\item{minRL}{Minimum value for the product of the set sizes for R and L. Default is 25.}

\item{X}{Optional. Specify a region of the chromosome to calculate \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} for in the format \code{c(startposition, endposition)}. The start position and the end position should be within the extremes of the positions given in the \code{pos} vector. If not supplied, the function will calculate \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} for every SNP in the \code{pos} vector.}
}
\value{
A list containing the SNP positions and the \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} values for those SNPs
}
\description{
Returns a \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} value for each SNP location supplied to the function, based on
the expected \eqn{r^2} values given an LD profile and genetic distances.
For more information about the \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} statistic please see Jacobs (2016).
The \eqn{Z_{\beta}^{BetaCDF}}{Zbeta} statistic is defined as:
\deqn{Z_{\beta}^{BetaCDF}=\frac{\sum_{i \in L,j \in R}\frac{B(r^2_{i,j};a,b)}{B(a,b)}}{|L||R|}}
where \code{|L|} and \code{|R|} are the number of SNPs to the left and right of the current locus within the given window \code{ws}, \eqn{r^2}{r^2} is equal to
the squared correlation between a pair of SNPs, and \eqn{\frac{B(r^2_{i,j};a,b)}{B(a,b)}} is the cumulative distribution function for the Beta distribution given
the estimated a and b parameters from the LD profile.
}
\details{
The LD profile describes the expected correlation between SNPs at a given genetic distance, generated using simulations or
real data. Care should be taken to utilise an LD profile which is representative of the population in question. The LD
profile should consist of evenly-sized bins of distances (for example 0.00001 cM per bin), where the value given is the (inclusive) lower
bound of the bin.
}
\references{
Jacobs, G.S., T.J. Sluckin, and T. Kivisild, \emph{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps.} Genetics, 2016. \strong{203}(4): p. 1807
}
