% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Runs the LR function}
\usage{
LR(pos, ws, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{ws}{The window size which the \code{LR} statistic will be calculated over. This should be on the same scale as the \code{pos} vector.}

\item{X}{Optional. Specify a region of the chromosome to calculate LR for in the format \code{c(startposition, endposition)}. The start position and the end position should be within the extremes of the positions given in the \code{pos} vector. If not supplied, the function will calculate LR for every SNP in the \code{pos} vector.}
}
\value{
A list containing the SNP positions and the \code{LR} values for those SNPs
}
\description{
Returns the \code{|L||R|} value for each SNP location supplied to the function.
For more information about the \code{|L||R|} diversity statistic please see Jacobs (2016).
}
\references{
Jacobs, G.S., T.J. Sluckin, and T. Kivisild, \emph{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps.} Genetics, 2016. \strong{203}(4): p. 1807
}
