% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{LR}
\alias{LR}
\title{Runs the LR function}
\usage{
LR(pos, x, ws, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{x}{A matrix of SNP values. Columns represent chromosomes, rows are SNP locations. Hence, the number of rows should equal the length of the `pos` vector. SNPs should all be biallelic.}

\item{ws}{The window size which the `LR` statistic will be calculated over. This should be on the same scale as the `pos` vector.}

\item{X}{Optional. Specify a region of the chromosome to calculate LR for in the format `c(startposition, endposition)`. The start position and the end position should be within the extremes of the positions given in the `pos` vector. If not supplied, the function will calculate LR for every SNP in the `pos` vector.}
}
\value{
A data frame containing the SNP positions and the LR values for those SNPs
}
\description{
Returns the |L||R| value for each SNP location supplied to the function.
For more information about the |L||R| diveristy statistic please see Jacobs (2016).
}
