% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zalpha.R
\name{Zalpha}
\alias{Zalpha}
\title{Runs the Zalpha function}
\usage{
Zalpha(pos, x, ws, minRandL = 4, minRL = 25, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{x}{A matrix of SNP values. Columns represent chromosomes, rows are SNP locations. Hence, the number of rows should equal the length of the \code{pos} vector. SNPs should all be biallelic.}

\item{ws}{The window size which the \eqn{Z_{\alpha}}{Zalpha} statistic will be calculated over. This should be on the same scale as the \code{pos} vector.}

\item{minRandL}{Minimum number of SNPs in each set R and L for the statistic to be calculated. Default is 4.}

\item{minRL}{Minimum value for the product of the set sizes for R and L. Default is 25.}

\item{X}{Optional. Specify a region of the chromosome to calculate \eqn{Z_{\alpha}}{Zalpha} for in the format \code{c(startposition, endposition)}. The start position and the end position should be within the extremes of the positions given in the \code{pos} vector. If not supplied, the function will calculate \eqn{Z_{\alpha}}{Zalpha} for every SNP in the \code{pos} vector.}
}
\value{
A list containing the SNP positions and the \eqn{Z_{\alpha}}{Zalpha} values for those SNPs
}
\description{
Returns a \eqn{Z_{\alpha}}{Zalpha} value for each SNP location supplied to the function.
For more information about the \eqn{Z_{\alpha}}{Zalpha} statistic please see Jacobs (2016).
The \eqn{Z_{\alpha}}{Zalpha} statistic is defined as:
\deqn{Z_{\alpha}=\frac{{|L| \choose 2}^{-1}\sum_{i,j \in L}r^2_{i,j} + {|R| \choose 2}^{-1}\sum_{i,j \in L}r^2_{i,j}}{2}}
where \code{|L|} and \code{|R|} are the number of SNPs to the left and right of the current locus within the given window \code{ws}, and \eqn{r^2}{r^2} is equal to the squared correlation between a pair of SNPs
}
\examples{
## load the snps example dataset
data(snps)
## run Zalpha over all the SNPs with a window size of 3000 bp
Zalpha(snps$positions,as.matrix(snps[,3:12]),3000)
## only return results for SNPs between locations 600 and 1500 bp
Zalpha(snps$positions,as.matrix(snps[,3:12]),3000,X=c(600,1500))

}
\references{
Jacobs, G.S., T.J. Sluckin, and T. Kivisild, \emph{Refining the Use of Linkage Disequilibrium as a Robust Signature of Selective Sweeps.} Genetics, 2016. \strong{203}(4): p. 1807
}
