% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zbeta.R
\name{Zbeta}
\alias{Zbeta}
\title{Runs the Zbeta function}
\usage{
Zbeta(pos, x, ws, minRandL = 4, minRL = 25, X = NULL)
}
\arguments{
\item{pos}{A numeric vector of SNP locations}

\item{x}{A matrix of SNP values. Columns represent chromosomes, rows are SNP locations. Hence, the number of rows should equal the length of the `pos` vector. SNPs should all be biallelic.}

\item{ws}{The window size which the `Zbeta` statistic will be calculated over. This should be on the same scale as the `pos` vector.}

\item{minRandL}{Minimum number of SNPs in each set R and L for the statistic to be calculated. Default is 4.}

\item{minRL}{Minimum value for the product of the set sizes for R and L. Default is 25.}

\item{X}{Optional. Specify a region of the chromosome to calculate Zbeta for in the format `c(startposition, endposition)`. The start position and the end position should be within the extremes of the positions given in the `pos` vector. If not supplied, the function will calculate Zbeta for every SNP in the `pos` vector.}
}
\value{
A data frame containing the SNP positions and the Zbeta values for those SNPs
}
\description{
Returns a Zbeta value for each SNP location supplied to the function.
For more information about the Zbeta statistic please see Jacobs (2016).
}
